requirements.txt
Django==5.2.1
djangorestframework==3.16.0
django-cors-headers==4.3.1
channels==4.2.2
channels-redis==4.2.1
celery==5.5.3
redis==5.0.1
ffmpeg-python==0.2.0
requests==2.31.0
python-decouple==3.8
psycopg2-binary==2.9.9
